feat: Moves Profile to Single Page App (SPA)
