+++ superset-frontend/src/features/home/RightMenu.tsx
@@ -474,7 +474,7 @@ const RightMenu = ({
             <Menu.ItemGroup key="user-section" title={t('User')}>
               {navbarRight.user_profile_url && (
                 <Menu.Item key="profile">
-                  <a href={navbarRight.user_profile_url}>{t('Profile')}</a>
+                  <Link to={navbarRight.user_profile_url}>{t('Profile')}</Link>
                 </Menu.Item>
               )}
               {navbarRight.user_info_url && (
