+++ superset-frontend/src/features/profile/Favorites.test.tsx
@@ -20,8 +20,8 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import thunk from 'redux-thunk';
 import configureStore from 'redux-mock-store';
-import Favorites from 'src/profile/components/Favorites';
 import TableLoader from 'src/components/TableLoader';
+import Favorites from './Favorites';
 
 import { user } from './fixtures';
