+++ superset-frontend/src/views/routes.tsx
@@ -119,6 +119,10 @@ const RowLevelSecurityList = lazy(
     ),
 );
 
+const Profile = lazy(
+  () => import(/* webpackChunkName: "Profile" */ 'src/pages/Profile'),
+);
+
 type Routes = {
   path: string;
   Component: React.ComponentType;
@@ -216,6 +220,10 @@ export const routes: Routes = [
   {
     path: '/rowlevelsecurity/list',
     Component: RowLevelSecurityList,
+  },
+  {
+    path: '/profile',
+    Component: Profile,
   },
 ];
